Added notification models, tables, and helpers

+++ db/migrate/20180723201606_remove_index_votes_on_user_id_and_post_id.rb
@@ -1,5 +1,5 @@
 class RemoveIndexVotesOnUserIdAndPostId < ActiveRecord::Migration
   def change
-    remove_index :votes, [:user_id, :post_id]
+    remove_index :votes, [:user_id, :post_id] if index_exists?(:votes, [:user_id, :post_id])
   end
 end
